Explode routing file
