+++ src/app/Routing/Route.ts
@@ -1,0 +1,8 @@
+import {RequestHandler} from 'express';
+
+export interface IRoute {
+    uri: string;
+    methods?: string[];
+    middlewares?: RequestHandler[];
+    action: string;
+}
