+++ src/app/Routing/Socket.ts
@@ -1,0 +1,6 @@
+import {RequestHandler} from 'express';
+
+export interface ISocket {
+    event: string;
+    action: string;
+}
