+++ src/app/index.ts
@@ -1,3 +1,3 @@
-export * from './Routing/Router';
+export * from './Routing';
 export * from './Utils/ModuleProvider';
 export * from './Kernel';
